Capture and Log Runtime Performance Metrics for Use in Optimization
Motivation
----------

https://stitchdata.atlassian.net/browse/SUP-279: Zendesk `tickets`
replication is too slow for us to ever finish

We believe that the substream strategy in this tap leads clients to be
unable to replicate their data fast enough to keep up with time. We think
this is intrinsic to the way Zendesk exposes their data but for the time
being we're going to start capturing and logging rates to facilitate and
measure optimization later.